# docs/swagger.yaml
basePath: /
definitions:
  structures.Item:
    properties:
      id:
        description: Unique ID
        type: string
      name:
        description: Object's name
        type: string
      price:
        description: Price
        type: number
    type: object
  structures.Response:
    properties:
      detail:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: CRUD API with default library
  title: CRUD Go API
  version: "1.0"
paths:
  /items:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/structures.Item'
            type: object
      summary: Return list of all Items.
      tags:
      - Items
    post:
      consumes:
      - application/json
      parameters:
      - description: New Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/structures.Item'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/structures.Item'
      summary: Create a new Item object.
      tags:
      - Items
  /items/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Object's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structures.Response'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
      summary: delete Item by ID.
      tags:
      - Items
    get:
      consumes:
      - application/json
      parameters:
      - description: Object's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structures.Item'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
      summary: get Item by ID.
      tags:
      - Items
swagger: "2.0"